requirements.txt
telethon-sync
cryptg
python-telegram-bot

requests
flask

minio

sqlalchemy
mysql-connector-python
redis

raven

gevent

twilio
